% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_openTab}
\alias{add_openTab}
\title{Convert a vector of customer names into a vector of customer page links.}
\usage{
add_openTab(to_menu, to_tab = "", modify_vector, display_vector = "", from_tab)
}
\arguments{
\item{to_menu}{A character value specifying the menu name to jump to.}

\item{to_tab}{A character value specifying the tab name to jump to (optional).}

\item{modify_vector}{A character vector of customer names to create links.}

\item{display_vector}{A character vector of customer names for display.}

\item{from_tab}{A character value specifying the current page.}
}
\value{
A character vector of customer page links.
}
\description{
Add html elements to a character vector to convert it into clickable links.
}
\examples{
add_openTab(to_menu = "customer", modify_vector = c("Customer A", "Customer B"),
from_tab = "search")
}
