% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Create a data.frame from pre-defined variable groups in a grouped.df object}
\usage{
compose(grouped.df, groups = NULL, names = FALSE)
}
\arguments{
\item{grouped.df}{A grouped.df object.}

\item{groups}{A character vector containing the variable group names.}

\item{names}{If TRUE, appends the variable group name as a prefix to the individual variables in that group.}
}
\value{
A data.frame.
}
\description{
Returns a data.frame containing the variables of a given set of groups.
}
\examples{
# create a new data.frame from a grouped.df
clin <- compose(g, groups=c('phenotype','medication'))
head(clin)

}
