% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate several data.frames together}
\usage{
consolidate(..., by = NULL, all.x = TRUE, all.y = FALSE)
}
\arguments{
\item{...}{Multiple data.frames.}

\item{by}{The variable by which to merge.}

\item{all.x}{If TRUE, merges such that all IDs in the lefthand side data.frame are kept.}

\item{all.y}{If TRUE, merges such that all IDs across all data.frames are kept.}
}
\value{
A data.frame.
}
\description{
A chained merge command of data.frames that share an ID variable.
}
\examples{
pheno <- data.frame(id=id,
age=rnorm(100,65,5.8),
male=rbinom(100,1,0.5),
bmi=rnorm(100,33,5),
cancer=rbinom(100,1,0.15),
afib=rbinom(100,1,0.25),
ckd=rbinom(100,1,0.35),
chf=rbinom(100,1,0.20),
copd=rbinom(100,1,0.3))

genotype <- data.frame(cbind(id,matrix(sample(2,5e6,replace=TRUE),nrow=100,ncol=5e4)))
colnames(genotype) <- c('id',paste0('gene_',seq(1,5e4)))

protein <- data.frame(cbind(id,matrix(rnorm(2e5,0,0.35),nrow=100,ncol=2e3)))
colnames(protein) <- c('id',paste0('protein_',seq(1,2e3)))
# now consolidate
highdim_df <- consolidate(pheno,medi,genotype,protein)
dim(highdim_df)
}
