% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{custom_md5_hash}
\alias{custom_md5_hash}
\title{Fast md5 check}
\usage{
custom_md5_hash(path_files)
}
\arguments{
\item{path_files}{a character vector containing absolute file paths.}
}
\value{
a character vector containing md5 values.
}
\description{
md5 values are computed based on combination of file path, file size and file modified time.
}
\examples{
custom_md5_hash(path_files = list.files(full.names = TRUE))
}
