% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dcf_valuation}
\alias{dcf_valuation}
\title{Calculate DCF Value}
\usage{
dcf_valuation(
  fcf,
  growth_rate,
  growth_years = 10,
  terminal_rate = 0.03,
  discount_rate = 0.15,
  shares = 1,
  excess_cash = 0
)
}
\arguments{
\item{fcf}{A numeric vector containing Free Cash Flow.}

\item{growth_rate}{A numeric vector containing Growth Rate.}

\item{growth_years}{A numeric vector containing Years of Growth. Default 10.}

\item{terminal_rate}{A numeric vector containing Terminal Rate. Default 0.03.}

\item{discount_rate}{A numeric vector containing Discount Rate. Default 0.15.}

\item{shares}{A numeric vector containing number of Outstanding Shares. Default 1.}

\item{excess_cash}{A numeric vector containing Excess Cash.}
}
\value{
A numeric vector with DCF Value.
}
\description{
Calculate DCF Value
}
\examples{
dcf_valuation(fcf = 1000, growth_rate = 0.10)
}
