% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expected_return}
\alias{expected_return}
\title{Calculate Expected Return from investment}
\usage{
expected_return(
  fcf,
  growth_rate,
  growth_years = 10,
  terminal_rate = 0.03,
  discount_rate = 0.15,
  shares,
  price,
  dcf
)
}
\arguments{
\item{fcf}{A numeric vector containing Free Cash Flow.}

\item{growth_rate}{A numeric vector containing Growth Rate.}

\item{growth_years}{A numeric vector containing Years of Growth. Default 10.}

\item{terminal_rate}{A numeric vector containing Terminal Rate. Default 0.03.}

\item{discount_rate}{A numeric vector containing Discount Rate. Default 0.15.}

\item{shares}{A numeric vector containing number of Outstanding Shares.}

\item{price}{A numeric vector containing current share price.}

\item{dcf}{A numeric vector containing DCF Valuation.}
}
\value{
A numeric vector with Expected Return.
}
\description{
Calculate Expected Return from investment
}
\examples{
expected_return(fcf = 100000000, growth_rate = 0.10, shares = 100000, price = 100, dcf = 10)
}
