% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{feather_to_DT}
\alias{feather_to_DT}
\title{Read a feather file and return a data.table object}
\usage{
feather_to_DT(path, columns = NULL)
}
\arguments{
\item{path}{A character vector containing an absolute path to a feather file.}

\item{columns}{Columns to read (names or indexes). Default: Read all columns.}
}
\value{
A data.table.
}
\description{
Read a feather file and return a data.table object
}
