% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_number}
\alias{format_number}
\title{Format a numeric vector into K, M, B, T}
\usage{
format_number(num_col, digits = 0)
}
\arguments{
\item{num_col}{A numeric vector.}

\item{digits}{A numeric value representing number of decimal digits.}
}
\value{
A numeric vector with KMBT format.
}
\description{
Format a numeric vector into K, M, B, T
}
\examples{
format_number(c(10, 42385, 8921478432, NA))
}
