% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply.R
\name{gapply}
\alias{gapply}
\title{An apply function for grouped.df objects}
\usage{
gapply(grouped.df, groupname, fun)
}
\arguments{
\item{grouped.df}{A grouped.df object.}

\item{groupname}{A character string referring to a variable group in a grouped.df object.}

\item{fun}{The function to be applied to all variables in the specified group.}
}
\value{
An element.df object after applying the function.
}
\description{
Applies a function to a variable group in a grouped.df object.
}
\examples{
# convert all variables in the group 'proteome' to numeric
head(sapply(g$proteome,class))
g$proteome <- gapply(g,'proteome',as.numeric)
head(sapply(g$proteome,class))
}
