% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{\%by\%}
\alias{\%by\%}
\title{By - an operator to apply a function stratified by a variable}
\usage{
fun \%by\% by
}
\arguments{
\item{fun}{A function.}

\item{by}{A variable to perform the function stratified by.}
}
\value{
The result of the function.
}
\description{
Does the same as aggregate() in an intuitive manner.
}
\examples{
mean(iris$Sepal.Length) \%by\% iris$Species
}
