% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{\%where\%}
\alias{\%where\%}
\title{Where - a subsetting operator}
\usage{
x \%where\% subset
}
\arguments{
\item{x}{A data.frame or matrix}
}
\value{
A subset of the original data.frame or matrix.
}
\description{
Does the same as subset().
}
\examples{
iris \%where\% (Sepal.Length > 6)
}
