% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_formula.R
\name{grouped_formula}
\alias{grouped_formula}
\title{A formula function for grouped.df objects}
\usage{
grouped_formula(grouped.df, gform, scale = FALSE)
}
\arguments{
\item{grouped.df}{A grouped.df object.}

\item{gform}{A formula using variable group names instead of individual variable names.}

\item{scale}{Scales all numeric variables in the formula.}
}
\value{
A variable-level formula.
}
\description{
Easily write a variable-level formula by using group-based shorthand notation.
}
\examples{
grouped_formula(g,outcome~phenotype+medication)
grouped_formula(g,outcome~phenotype+medication+protein_555+gene_2450)
}
