% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{log_messages}
\alias{log_messages}
\title{Create/update log file with the logged data}
\usage{
log_messages(message, file_path)
}
\arguments{
\item{message}{A character string representing the message to log.}

\item{file_path}{A character string representing the path of the log file. Do not include file extension.}
}
\value{
A logical value TRUE if the operation is successful.
}
\description{
Log file is created or updated with the latest message along with the log timestamp.
Enhanced for multi-threading if multiple threads are trying to log messages at the same time.
}
