% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mappings}
\alias{mappings}
\title{Show all grouped mappings for a data.frame in the global environment}
\usage{
mappings(originaldf)
}
\arguments{
\item{originaldf}{A data.frame in the global environment.}
}
\description{
Identifies all grouped.dfs corresponding to a given data.frame in the global environment.
}
\examples{
mappings(highdim_df)

}
