% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.grouped.df.R
\name{plot.grouped.df}
\alias{plot.grouped.df}
\title{A plotting function for grouped.df objects}
\usage{
\method{plot}{grouped.df}(grouped.df)
}
\value{
A barchart.
}
\description{
Creates a barchart of the variable groups in a grouped.df.
}
