% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{+.element.df}
\alias{+.element.df}
\title{An operator to quickly compose data.frames from predefined variable groups}
\usage{
\method{+}{element.df}(...)
}
\value{
A data.frame.
}
\description{
Does the same as compose() in an intuitive manner.
}
\examples{
clin <- g$phenotype + g$medication
head(clin)
}
