% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_md5_folder}
\alias{read_md5_folder}
\title{Generate md5 hash of the files in a given folder}
\usage{
read_md5_folder(folder_path)
}
\arguments{
\item{folder_path}{A character string representing an absolute folder path.}
}
\value{
A character vector with md5 hash values of all the files within a given folder.
}
\description{
Generates the md5 hash of the files in a given folder. Note that certain file types are ignored such as those end with .trc, .xlsm, .log and those containing ~$.
custom_md5_hash from package "simplify" is used to generate md5 hash.
}
\examples{
read_md5_folder(folder_path = getwd())
}
