% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize the first letter of all words}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{A character value (string) preferably lowercase string.}
}
\value{
A character value (string) with first letter of each word in capital.
}
\description{
Make first letter of every word uppercase. It is advised to use a string with all lower case as input.
}
\examples{
simpleCap("watch the magic")
}
