% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.grouped.df.R
\name{summary.grouped.df}
\alias{summary.grouped.df}
\title{A summary function for grouped.df objects}
\usage{
\method{summary}{grouped.df}(grouped.df)
}
\description{
Returns an overview of the variable groups in a grouped.df object.
}
