% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{unescape_html}
\alias{unescape_html}
\title{Unescape html characters}
\usage{
unescape_html(str)
}
\arguments{
\item{str}{A character string containing HTML characers}
}
\value{
An unescaped character string
}
\description{
Unescape html characters
}
\examples{
unescape_html("&euro; 2.99")
}
