% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{update_md5_file}
\alias{update_md5_file}
\title{Update the md5 hash once all the files in a given folder have processed}
\usage{
update_md5_file(md5_data, folder_path)
}
\arguments{
\item{md5_data}{A character vector representing md5 hash values of the files within the given folder.}

\item{folder_path}{A character string representing an absolute folder path.}
}
\value{
A logical value TRUE if the operation is successful.
}
\description{
A csv file with the same name as the folder name is created/updated with new md5 values. This file with md5 hash values and the given folder has same parent directory.
}
