% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{update_output_file}
\alias{update_output_file}
\title{Update/create a file with the given data at a given path}
\usage{
update_output_file(data, file_path, max_limit = 3)
}
\arguments{
\item{data}{A data.frame, tibble or data.table.}

\item{file_path}{A character string representing the path of the output file. Do not include file extension.}

\item{max_limit}{A numeric value representing the number of most recent back up files to keep. Default is 3.}
}
\value{
A logical value TRUE if the operation is successful.
}
\description{
Updates data in a file based on the given path. Backs up the old file if the file with the same name is already present.
Appends the file name with the timestamp for the back up files.
}
