% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{feather_to_DT}
\alias{feather_to_DT}
\title{Read a feather file and return a data.table object}
\usage{
feather_to_DT(file_path)
}
\arguments{
\item{file_path}{A character vector containing an absolute path to a feather file.}
}
\value{
A data.table.
}
\description{
Read a feather file and return a data.table object
}
