% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{file_append}
\alias{file_append}
\title{Append multiple files in a given folder into one}
\usage{
file_append(folder_path, sheet = NULL)
}
\arguments{
\item{folder_path}{A character string representing an absolute folder path.}

\item{sheet}{A character string representing name of the excel sheet.}
}
\value{
A data.table representing the combined data of all the files within that given folder.
}
\description{
Appends all the files in a folder. Files have to be either csv, xlsx or txt. Note that certain file types are ignored such as those end with .trc, .xlsm, .log and those containing ~$.
custom_md5_hash from package "simplify" is used to generate md5 hash.
}
